requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
bs4==0.0.1
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.7
django-adminlte-3==0.1.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-google-maps==0.13.0
django-material==1.10.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-gis==1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
env==0.1.0
et-xmlfile==1.1.0
Faker==10.0.0
filelock==3.3.0
GDAL @ file:///F:/thekreativeideas/geolocation/GDAL-3.2.3-cp39-cp39-win_amd64.whl
geographiclib==1.52
geopy==2.2.0
ghostscript==0.7
googlemaps==4.6.0
greenlet==1.1.2
gunicorn==20.1.0
httpie==2.6.0
idna==3.2
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
module==0.0.4
mysql-connector-python==8.0.29
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.3
opencv-python==4.5.4.58
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pdfminer==20191125
pdfminer.six==20211012
pdftables==0.0.3
pdftopng==0.2.3
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
plyer==2.0.0
postgres==4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.4
psycopg2==2.9.1
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.20
pycryptodome==3.11.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==3.0.0
PyPDF2==1.26.0
PyQRCode==1.2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pywin32==302
pywinpty==1.1.4
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
query==0.1.4
razorpay==1.3.0
reg==0.12
repoze.lru==0.7
requests==2.26.0
requests-toolbelt==0.9.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
sqlparse==0.4.2
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
virtualenv==20.8.1
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
widgetsnbextension==3.5.1
